add last_evaluated_key attribute on table generator move type-specific item_object_hook() into types.py fix bug with max_results=0 yielding all results PEP8 cleanups
