+++ boto/dynamodb/layer2.py
@@ -20,38 +20,13 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 #
-import base64
-
 from boto.dynamodb.layer1 import Layer1
 from boto.dynamodb.table import Table
 from boto.dynamodb.schema import Schema
 from boto.dynamodb.item import Item
 from boto.dynamodb.batch import BatchList, BatchWriteList
 from boto.dynamodb.types import get_dynamodb_type, dynamize_value, \
-        convert_num, convert_binary
-
-
-def item_object_hook(dct):
-    """
-    A custom object hook for use when decoding JSON item bodys.
-    This hook will transform Amazon DynamoDB JSON responses to something
-    that maps directly to native Python types.
-    """
-    if len(dct.keys()) > 1:
-        return dct
-    if 'S' in dct:
-        return dct['S']
-    if 'N' in dct:
-        return convert_num(dct['N'])
-    if 'SS' in dct:
-        return set(dct['SS'])
-    if 'NS' in dct:
-        return set(map(convert_num, dct['NS']))
-    if 'B' in dct:
-        return convert_binary(dct['B'])
-    if 'BS' in dct:
-        return set(map(convert_binary, dct['BS']))
-    return dct
+    item_object_hook
 
 
 def table_generator(tgen):
@@ -61,27 +36,33 @@ def table_generator(tgen):
     :class:`boto.dynamodb.layer2.TableGenerator` and is not intended
     to be used outside of that context.
     """
-    response = True
     n = 0
-    while response:
-        if tgen.max_results and n == tgen.max_results:
+    while True:
+        response = tgen.callable(**tgen.kwargs)
+        if not response:
             break
-        if response is True:
-            pass
-        elif 'LastEvaluatedKey' in response:
+        tgen.consumed_units += response.get('ConsumedCapacityUnits', 0)
+        # at the expense of a possibly gratuitous dynamize, ensure that
+        # early generator termination won't result in bad LEK values
+        if 'LastEvaluatedKey' in response:
             lek = response['LastEvaluatedKey']
             esk = tgen.table.layer2.dynamize_last_evaluated_key(lek)
             tgen.kwargs['exclusive_start_key'] = esk
+            lektuple = (lek['HashKeyElement'],)
+            if 'RangeKeyElement' in lek:
+                lektuple += (lek['RangeKeyElement'],)
+            tgen.last_evaluated_key = lektuple
         else:
-            break
-        response = tgen.callable(**tgen.kwargs)
-        if 'ConsumedCapacityUnits' in response:
-            tgen.consumed_units += response['ConsumedCapacityUnits']
+            tgen.last_evaluated_key = None
         for item in response['Items']:
-            if tgen.max_results and n == tgen.max_results:
+            if tgen.max_results is not None and n == tgen.max_results:
                 break
             yield tgen.item_class(tgen.table, attrs=item)
             n += 1
+        else:
+            if tgen.last_evaluated_key is not None:
+                continue
+        break
 
 
 class TableGenerator:
@@ -94,6 +75,10 @@ class TableGenerator:
     :ivar consumed_units: An integer that holds the number of
         ConsumedCapacityUnits accumulated thus far for this
         generator.
+
+    :ivar last_evaluated_key: A sequence representing the key(s)
+        of the item last evaluated, or None if no additional
+        results are available.
     """
 
     def __init__(self, table, callable, max_results, item_class, kwargs):
@@ -103,6 +88,7 @@ class TableGenerator:
         self.item_class = item_class
         self.kwargs = kwargs
         self.consumed_units = 0
+        self.last_evaluated_key = None
 
     def __iter__(self):
         return table_generator(self)
